+++ lib/mapster.dart
@@ -2,5 +2,5 @@ library mapster;
 
 export 'src/mappers/mappers.dart' hide getSource;
 export 'src/mapster/exceptions/exceptions.dart';
-export 'src/mapster/mapper_creator.dart';
+export 'src/mapster/mapper_meta.dart';
 export 'src/mapster/mapster.dart';