+++ lib/src/mapster/mapper_creator.dart
@@ -1,4 +1,0 @@
-import '../mappers/mapper.dart';
-
-/// Callback for creating [Mapper].
-typedef MapperCreator<M extends Mapper> = M Function(MapperInput input);