+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Change signature of `register`, `registerAll` in `Mapster`
 - Add `MapperInput`, `MapperCreator`
 - Fix mappers (from `FourSourcesMapper` to `NineSourcesMapper`)
+- Remove `to`, `fromTypes` in `Mapper`
 
 ## 0.1.0